+++ src/main/kotlin/com/github/rafaelldi/tyeplugin/actions/TyeBuildAction.kt
@@ -3,27 +3,31 @@
 package com.github.rafaelldi.tyeplugin.actions
 
 import com.github.rafaelldi.tyeplugin.runtimes.TyeApplicationRuntime
-import com.github.rafaelldi.tyeplugin.services.TyeApplicationManager
+import com.github.rafaelldi.tyeplugin.services.TyeCliService
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.components.service
 import com.intellij.openapi.progress.runBackgroundableTask
 import com.intellij.openapi.project.DumbAwareAction
 import com.intellij.remoteServer.util.ApplicationActionUtils
 
-class TyeApplicationShutdownAction : DumbAwareAction() {
+class TyeBuildAction : DumbAwareAction() {
     override fun actionPerformed(e: AnActionEvent) {
+        val project = e.project ?: return
         val runtime = ApplicationActionUtils.getApplicationRuntime(e, TyeApplicationRuntime::class.java) ?: return
-        val tyeApplicationManager: TyeApplicationManager = service()
+        val sourcePath = runtime.getSourceFile()?.path ?: return
+        val tyeCliService = project.service<TyeCliService>()
 
-        runBackgroundableTask("Tye shutdown", e.project) {
+        runBackgroundableTask("Tye build", e.project) {
             it.isIndeterminate = true
-            it.text = "Tye application is stopping"
-            tyeApplicationManager.shutdownApplication(runtime.host)
+            it.text = "Tye build command is running"
+            tyeCliService.runBuildCommand(sourcePath, project.basePath)
         }
     }
 
     override fun update(e: AnActionEvent) {
+        val project = e.project
         val runtime = ApplicationActionUtils.getApplicationRuntime(e, TyeApplicationRuntime::class.java)
-        e.presentation.isEnabledAndVisible = runtime != null
+        val sourcePath = runtime?.getSourceFile()?.path
+        e.presentation.isEnabledAndVisible = project != null && sourcePath != null
     }
 }